Extract ensure mesh functions and constants into own file
