+++ edit.js
@@ -10,11 +10,11 @@ import {wireframeMaterial, getWireframeMesh, meshIdToArray, decorateRaycastMesh,
 import {makeLineMesh, makeTeleportMesh} from './teleport.js';
 
 import {handleUrl, addPackage} from './edit/utils.js';
-import targetMeshGeometry from './edit/targetMeshGeometry.js';
 import attachEventListeners from './edit/classlistHandlers.js';
 import dragHandlers from './edit/dragHandlers.js';
 import {updateWorldSaveButton, worldHandlers} from './edit/worlds.js';
-import {apiHost, packagesEndpoint, contract} from './edit/constants.js';
+import {apiHost, packagesEndpoint, contract, loadMeshMaterial} from './edit/constants.js';
+import {ensureLoadMesh, ensurePlaceholdMesh, ensureVolumeMesh} from './edit/meshes.js';
 import {
   packagesSubpage, inventorySubpage, avatarSubpage, avatarSubpageContent,
   tabs, tabContents, inventorySubtabContent,
@@ -50,54 +50,6 @@ let jumpState = null;
 const _resetKeys = () => {
   for (const k in keys) {
     keys[k] = false;
-  }
-};
-
-const targetVsh = `
-  #define M_PI 3.1415926535897932384626433832795
-  uniform float uTime;
-  // varying vec2 vUv;
-  void main() {
-    float f = 1.0 + pow(sin(uTime * M_PI), 0.5) * 0.2;
-    gl_Position = projectionMatrix * modelViewMatrix * vec4(position * f, 1.);
-  }
-`;
-const targetFsh = `
-  uniform float uHighlight;
-  uniform float uTime;
-  void main() {
-    float f = max(1.0 - pow(uTime, 0.5), 0.1);
-    gl_FragColor = vec4(vec3(f * uHighlight), 1.0);
-  }
-`;
-const _makeTargetMesh = p => {
-  const geometry = targetMeshGeometry;
-  const material = new THREE.ShaderMaterial({
-    uniforms: {
-      uHighlight: {
-        type: 'f',
-        value: 0,
-      },
-      uTime: {
-        type: 'f',
-        value: 0,
-      },
-    },
-    vertexShader: targetVsh,
-    fragmentShader: targetFsh,
-    // transparent: true,
-  });
-  const mesh = new THREE.Mesh(geometry, material);
-  mesh.frustumCulled = false;
-  return mesh;
-};
-const _makeVolumeMesh = async p => {
-  const volumeMesh = await p.getVolumeMesh();
-  if (volumeMesh) {
-    volumeMesh.frustumCulled = false;
-    return volumeMesh;
-  } else {
-    return new THREE.Object3D();
   }
 };
 
@@ -679,92 +631,6 @@ window.addEventListener('mouseup', e => {
   lastTeleport = false;
 });
 
-const loadVsh = `
-  #define M_PI 3.1415926535897932384626433832795
-  uniform float uTime;
-
-  mat4 rotationMatrix(vec3 axis, float angle)
-  {
-      axis = normalize(axis);
-      float s = sin(angle);
-      float c = cos(angle);
-      float oc = 1.0 - c;
-
-      return mat4(oc * axis.x * axis.x + c,           oc * axis.x * axis.y - axis.z * s,  oc * axis.z * axis.x + axis.y * s,  0.0,
-                  oc * axis.x * axis.y + axis.z * s,  oc * axis.y * axis.y + c,           oc * axis.y * axis.z - axis.x * s,  0.0,
-                  oc * axis.z * axis.x - axis.y * s,  oc * axis.y * axis.z + axis.x * s,  oc * axis.z * axis.z + c,           0.0,
-                  0.0,                                0.0,                                0.0,                                1.0);
-  }
-
-  void main() {
-    // float f = 1.0 + pow(sin(uTime * M_PI), 0.5) * 0.2;
-    gl_Position = projectionMatrix * modelViewMatrix * rotationMatrix(vec3(0, 0, 1), -uTime * M_PI * 2.0) * vec4(position, 1.);
-  }
-`;
-const loadFsh = `
-  uniform float uHighlight;
-  uniform float uTime;
-  void main() {
-    float f = 1.0 + max(1.0 - uTime, 0.0);
-    gl_FragColor = vec4(vec3(${new THREE.Color(0xf4511e).toArray().join(', ')}) * f, 1.0);
-  }
-`;
-const loadMeshMaterial = new THREE.ShaderMaterial({
-  uniforms: {
-    /* uHighlight: {
-      type: 'f',
-      value: 0,
-    }, */
-    uTime: {
-      type: 'f',
-      value: 0,
-    },
-  },
-  vertexShader: loadVsh,
-  fragmentShader: loadFsh,
-  side: THREE.DoubleSide,
-});
-const _makeLoadMesh = (() => {
-  const geometry = new THREE.RingBufferGeometry(0.05, 0.08, 128, 0, Math.PI / 2, Math.PI * 2 * 0.9);
-  // .applyMatrix4(new THREE.Matrix4().makeRotationFromQuaternion(new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(0, 1, 0), Math.PI)));
-  return () => {
-    const mesh = new THREE.Mesh(geometry, loadMeshMaterial);
-    // mesh.frustumCulled = false;
-    return mesh;
-  };
-})();
-const _ensureLoadMesh = p => {
-  if (!p.loadMesh) {
-    p.loadMesh = _makeLoadMesh();
-    p.loadMesh.matrix.copy(p.matrix).decompose(p.loadMesh.position, p.loadMesh.quaternion, p.loadMesh.scale);
-    scene.add(p.loadMesh);
-
-    p.waitForRun()
-      .then(() => {
-        p.loadMesh.visible = false;
-      });
-  }
-};
-const _ensurePlaceholdMesh = p => {
-  if (!p.placeholderBox) {
-    p.placeholderBox = _makeTargetMesh();
-    p.placeholderBox.package = p;
-    p.placeholderBox.matrix.copy(p.matrix).decompose(p.placeholderBox.position, p.placeholderBox.quaternion, p.placeholderBox.scale);
-    p.placeholderBox.visible = false;
-    scene.add(p.placeholderBox);
-  }
-};
-const _ensureVolumeMesh = async p => {
-  if (!p.volumeMesh) {
-    p.volumeMesh = await _makeVolumeMesh(p);
-    p.volumeMesh = getWireframeMesh(p.volumeMesh);
-    decorateRaycastMesh(p.volumeMesh, p.id);
-    p.volumeMesh.package = p;
-    p.volumeMesh.matrix.copy(p.matrix).decompose(p.volumeMesh.position, p.volumeMesh.quaternion, p.volumeMesh.scale);
-    p.volumeMesh.visible = false;
-    scene.add(p.volumeMesh);
-  }
-};
 shieldSlider.addEventListener('change', async e => {
   const newShieldLevel = parseInt(e.target.value, 10);
   const {packages} = pe;
@@ -820,9 +686,9 @@ const _packageadd = async e => {
     reason,
   } = e.data;
 
-  _ensureLoadMesh(p);
-  _ensurePlaceholdMesh(p);
-  await _ensureVolumeMesh(p);
+  ensureLoadMesh(p, scene);
+  ensurePlaceholdMesh(p, scene);
+  await ensureVolumeMesh(p, scene);
   _renderObjects();
 
   _bindObject(p);
